Moved AsyncSource to class. Removed isAsync checks,
