+++ src/sources/source.ts
@@ -7,18 +7,7 @@
  * @interface Source
  */
 export interface Source {
-    sourceType: string;
+    readonly sourceType: string;
     sourceId: string;
-    readonly isAsync: boolean;
     getAvatar(size?: number): string;
 }
-
-/**
- * Contract of all async sources.
- * Every async source must implement the processResponse method that extracts the avatar url from the data
- * @interface AsyncSource
- */
-export interface AsyncSource extends Source {
-    processResponse(data: any, size?: number): string;
-}
-
