+++ src/sources/twitter.ts
@@ -11,7 +11,6 @@ import { Source } from "./source";
  */
 export class Twitter implements Source {
     readonly sourceType: string = "TWITTER";
-    readonly isAsync = false;
 
     constructor(public sourceId: string) {
     }
